Middle way through Refactoring
