+++ headers/models/input_data_struct.h
@@ -1,0 +1,47 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   input_data_struct.h                                :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: schiper <schiper@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/04/29 14:42:48 by schiper           #+#    #+#             */
+/*   Updated: 2025/04/29 16:03:14 by schiper          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#if !defined(INPUT_DATA_STRUCT_H)
+# define INPUT_DATA_STRUCT_H
+
+#include "token_struct.h"
+
+typedef struct s_parse_params
+{
+	char	*input;
+	t_token	**tokens;
+	int		start;
+	int		end;
+	int		is_first;
+	char	**envp;
+	int		in_quote;
+	char	quote_char;
+	int		segment_start;
+	int		segment_end;
+	int		is_first_segment;
+	int		exit_status;
+}			t_parse_params;
+
+//??? Debatable if needed 
+typedef struct s_parse_state
+{
+	int				i;
+	int				start;
+	int				in_word;
+	int				error;
+	int				is_first_token;
+	int				in_quote;
+	char			quote_char;
+	t_token			**tokens;
+}					t_parse_state;
+
+#endif // INPUT_DATA_STRUCT_H
