+++ sources/graph/abstract_syntax_tree/graph_parsing.c
@@ -6,11 +6,11 @@
 /*   By: schiper <schiper@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/21 22:00:19 by schiper           #+#    #+#             */
-/*   Updated: 2025/04/28 15:53:19 by schiper          ###   ########.fr       */
+/*   Updated: 2025/04/29 14:18:28 by schiper          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "graph_functions.h"
+#include "abstract_syntax_tree.h"
 
 t_node	*parse_expression(t_parsed_data **tokens)
 {
