+++ sources/graph/abstract_syntax_tree/utils/token_type_to_node_type.c
@@ -6,11 +6,11 @@
 /*   By: schiper <schiper@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/25 17:32:31 by schiper           #+#    #+#             */
-/*   Updated: 2025/04/28 13:42:01 by schiper          ###   ########.fr       */
+/*   Updated: 2025/04/29 18:50:23 by schiper          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "graph_functions.h"
+#include "abstract_syntax_tree.h"
 
 t_node_type	token_type_to_node_type(const t_token_type *token_type)
 {
