+++ sources/input/parser/execution/parse_segment_data.c
