+++ sources/input/read_pipeline.c
@@ -6,12 +6,11 @@
 /*   By: schiper <schiper@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/26 14:39:17 by schiper           #+#    #+#             */
-/*   Updated: 2025/04/26 14:51:57 by schiper          ###   ########.fr       */
+/*   Updated: 2025/04/29 14:36:42 by schiper          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "input.h"
-#include "token_struct.h"
 #include "tokener.h"
 #include <stdio.h>
