+++ sources/minishell.c
@@ -6,11 +6,11 @@
 /*   By: schiper <schiper@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/10 14:31:05 by iatilla-          #+#    #+#             */
-/*   Updated: 2025/04/29 13:36:04 by schiper          ###   ########.fr       */
+/*   Updated: 2025/04/29 16:03:12 by schiper          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "graph_functions.h"
+#include "abstract_syntax_tree.h"
 #include "input.h"
 #include "tokener.h"
 
@@ -72,9 +72,11 @@ int	display_tokens(t_token *tokens)
 	if (!tokens)
 		return (1);
 	parsed_data = tokens_to_parsed_data(tokens);
-		copy_data = parsed_data;
-		ast_root = parse_expression(&copy_data);
-		dfs_walk(ast_root);
+    //Just for debug Purpose ATM
+	copy_data = parsed_data;
+	ast_root = parse_expression(&copy_data);
+	dfs_walk(ast_root);
+    //Just for debug Purpose ATM
 	if (parsed_data)
 	{
 		printf("Tokenization successful!\n");
