--- sources/tokener/execution/check_whitespace.c
@@ -1,37 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   check_whitespace.c                                 :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: iatilla- <iatilla-@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/04/23 20:19:09 by iatilla-          #+#    #+#             */
-/*   Updated: 2025/04/25 19:21:26 by iatilla-         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "tokener.h"
-
-/**
- * Handle whitespace in the input string
- * @param input: The input string
- * @param state: Parsing state
- * @param envp: Environment variables
- * @return: 1 if handled, 0 otherwise
- */
-int	handle_whitespace(char *input, t_parse_state *state, char **envp)
-{
-	if (ft_is_whitespace(input[state->i]))
-	{
-		if (state->in_word)
-		{
-			process_token(input, state, state->i, envp);
-			state->in_word = 0;
-		}
-		while (input[state->i] && ft_is_whitespace(input[state->i]))
-			state->i++;
-		state->start = state->i;
-		return (1);
-	}
-	return (0);
-}
