+++ sources/tokener/execution/handle_quotes/is_quote_closed.c
@@ -1,0 +1,30 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   is_quote_closed.c                                  :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: schiper <schiper@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/04/29 18:27:35 by schiper           #+#    #+#             */
+/*   Updated: 2025/04/29 18:27:53 by schiper          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "tokener.h"
+
+/**
+ * Check if a quote is closed properly
+ * @param str: The string to check
+ * @param start: Starting index of the quote
+ * @param quote_char: The quote character to match (' or ")
+ * @return: 1 if closed properly, 0 if unclosed
+ */
+int	is_quote_closed(char *str, int start, char quote_char)
+{
+	int	i;
+
+	i = start + 1;
+	while (str[i] && str[i] != quote_char)
+		i++;
+	return (str[i] == quote_char);
+}
